+++ R/rowCumsums.R
@@ -21,7 +21,7 @@
 #' @keywords array iteration univar
 #' @export
 rowCumsums <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCumsums, x, dim., rows, cols, TRUE, useNames)
 }
 
@@ -29,7 +29,7 @@ rowCumsums <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames
 #' @rdname rowCumsums
 #' @export
 colCumsums <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCumsums, x, dim., rows, cols, FALSE, useNames)
 }
 
@@ -37,7 +37,7 @@ colCumsums <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames
 #' @rdname rowCumsums
 #' @export
 rowCumprods <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCumprods, x, dim., rows, cols, TRUE, useNames)
 }
 
@@ -45,7 +45,7 @@ rowCumprods <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useName
 #' @rdname rowCumsums
 #' @export
 colCumprods <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCumprods, x, dim., rows, cols, FALSE, useNames)
 }
 
@@ -53,7 +53,7 @@ colCumprods <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useName
 #' @rdname rowCumsums
 #' @export
 rowCummins <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCummins, x, dim., rows, cols, TRUE, useNames)
 }
 
@@ -61,7 +61,7 @@ rowCummins <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames
 #' @rdname rowCumsums
 #' @export
 colCummins <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCummins, x, dim., rows, cols, FALSE, useNames)
 }
 
@@ -69,7 +69,7 @@ colCummins <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames
 #' @rdname rowCumsums
 #' @export
 rowCummaxs <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCummaxs, x, dim., rows, cols, TRUE, useNames)
 }
 
@@ -77,6 +77,6 @@ rowCummaxs <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames
 #' @rdname rowCumsums
 #' @export
 colCummaxs <- function(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowCummaxs, x, dim., rows, cols, FALSE, useNames)
 }