+++ R/rowDiffs.R
@@ -17,7 +17,7 @@
 #' @export
 rowDiffs <- function(x, rows = NULL, cols = NULL,
                      lag = 1L, differences = 1L, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowDiffs, x, dim., rows, cols, lag, differences, TRUE, useNames)
 }
 
@@ -25,6 +25,6 @@ rowDiffs <- function(x, rows = NULL, cols = NULL,
 #' @export
 colDiffs <- function(x, rows = NULL, cols = NULL,
                      lag = 1L, differences = 1L, dim. = dim(x), ..., useNames = TRUE) {
-  if (is.na(useNames)) deprecatedUseNamesNA()
+  if (is.na(useNames)) defunctUseNamesNA()
   .Call(C_rowDiffs, x, dim., rows, cols, lag, differences, FALSE, useNames)
 }