+++ R/sum2.R
@@ -6,16 +6,6 @@
 #' is faster and more memory efficient since it avoids the actual subsetting
 #' which requires copying of elements and garbage collection thereof.
 #'
-#' Furthermore, \code{sum2(x, mode = "double")} is equivalent to
-#' \code{sum(as.numeric(x))} and may therefore be used to avoid integer
-#' overflow(*), but at the same time is much more memory efficient that
-#' the regular \code{sum()} function when \code{x} is an
-#' \code{\link[base]{integer}} vector.
-#' 
-#' (*) \emph{In R (>= 3.5.0), \code{sum(x)} will no longer integer overflow
-#' and return \code{NA_integer_}. Instead it will return the correct sum in
-#' form of a double value.}
-#'
 #' @inheritParams rowAlls
 #' @inheritParams weightedMad
 #'