+++ tests/rowCumMinMaxs_subset.R
@@ -46,7 +46,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   for (rows in index_cases) {
     for (cols in index_cases) {
       # Check names attribute
-      for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
+      for (useNames in c(TRUE, FALSE)) {
         validateIndicesTestMatrix(x, rows, cols,
                                   ftest = rowCummins, fsure = rowCummins_R, useNames = useNames,verbose=TRUE)
         validateIndicesTestMatrix(x, rows, cols,