+++ tests/rowWeightedVars.R
@@ -24,12 +24,6 @@ x_est2 <- colWeightedVars(t(x), w = w, useNames = FALSE)
 stopifnot(all.equal(x_est1, x_est0))
 stopifnot(all.equal(x_est2, x_est0))
 x_est0 <- rowVars(x, useNames = TRUE)
-if (!matrixStats:::isUseNamesNADefunct()) {
-  x_est1 <- rowWeightedVars(x, w = w, useNames = NA)
-  x_est2 <- colWeightedVars(t(x), w = w, useNames = NA)
-  stopifnot(all.equal(x_est1, x_est0))
-  stopifnot(all.equal(x_est2, x_est0))
-}
 x_est1 <- rowWeightedVars(x, w = w, useNames = TRUE)
 x_est2 <- colWeightedVars(t(x), w = w, useNames = TRUE)
 stopifnot(all.equal(x_est1, x_est0))
@@ -54,12 +48,6 @@ stopifnot(all.equal(x_est1, x_est0))
 stopifnot(all.equal(x_est2, x_est0))
 x3 <- cbind(x, x, x)
 x_est0 <- rowVars(x3, useNames = TRUE)
-if (!matrixStats:::isUseNamesNADefunct()) {
-  x_est1 <- rowWeightedVars(x, w = w, useNames = NA)
-  x_est2 <- colWeightedVars(t(x), w = w, useNames = NA)
-  stopifnot(all.equal(x_est1, x_est0))
-  stopifnot(all.equal(x_est2, x_est0))
-}
 x_est1 <- rowWeightedVars(x, w = w, useNames = TRUE)
 x_est2 <- colWeightedVars(t(x), w = w, useNames = TRUE)
 stopifnot(all.equal(x_est1, x_est0))
@@ -82,12 +70,6 @@ x_est2 <- colWeightedVars(t(x), w = w, useNames = FALSE)
 stopifnot(all.equal(x_est1, x_est0))
 stopifnot(all.equal(x_est2, x_est0))
 x_est0 <- rowVars(x[, (w == 1), drop = FALSE], useNames = TRUE)
-if (!matrixStats:::isUseNamesNADefunct()) {
-  x_est1 <- rowWeightedVars(x, w = w, useNames = NA)
-  x_est2 <- colWeightedVars(t(x), w = w, useNames = NA)
-  stopifnot(all.equal(x_est1, x_est0))
-  stopifnot(all.equal(x_est2, x_est0))
-}
 x_est1 <- rowWeightedVars(x, w = w, useNames = TRUE)
 x_est2 <- colWeightedVars(t(x), w = w, useNames = TRUE)
 stopifnot(all.equal(x_est1, x_est0))
@@ -109,12 +91,6 @@ x_est2 <- colWeightedVars(t(x), w = w, useNames = FALSE)
 stopifnot(all.equal(x_est1, x_est0))
 stopifnot(all.equal(x_est2, x_est0))
 x_est0 <- rowVars(x[, (w == 1), drop = FALSE], useNames = TRUE)
-if (!matrixStats:::isUseNamesNADefunct()) {
-  x_est1 <- rowWeightedVars(x, w = w, useNames = NA)
-  x_est2 <- colWeightedVars(t(x), w = w, useNames = NA)
-  stopifnot(all.equal(x_est1, x_est0))
-  stopifnot(all.equal(x_est2, x_est0))
-}
 x_est1 <- rowWeightedVars(x, w = w, useNames = TRUE)
 x_est2 <- colWeightedVars(t(x), w = w, useNames = TRUE)
 stopifnot(all.equal(x_est1, x_est0))
@@ -136,12 +112,6 @@ x_est2 <- colWeightedVars(t(x), w = w, useNames = FALSE)
 stopifnot(all.equal(x_est1, x_est0))
 stopifnot(all.equal(x_est2, x_est0))
 x_est0 <- rowVars(x[, (w == 1), drop = FALSE], useNames = TRUE)
-if (!matrixStats:::isUseNamesNADefunct()) {
-  x_est1 <- rowWeightedVars(x, w = w, useNames = NA)
-  x_est2 <- colWeightedVars(t(x), w = w, useNames = NA)
-  stopifnot(all.equal(x_est1, x_est0))
-  stopifnot(all.equal(x_est2, x_est0))
-}
 x_est1 <- rowWeightedVars(x, w = w, useNames = TRUE)
 x_est2 <- colWeightedVars(t(x), w = w, useNames = TRUE)
 stopifnot(all.equal(x_est1, x_est0))
@@ -155,7 +125,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   if (setDimnames) dimnames(x) <- dimnames
   else dimnames(x) <- NULL    
   # Check names attribute
-  for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
+  for (useNames in c(TRUE, FALSE)) {
     x_est1 <- rowWeightedVars(x, w = w, useNames = useNames)
     print(x_est1)
     x_est2 <- colWeightedVars(t(x), w = w, useNames = useNames)
@@ -194,7 +164,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   if (setDimnames) dimnames(x) <- dimnames
   else dimnames(x) <- NULL    
   # Check names attribute
-  for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
+  for (useNames in c(TRUE, FALSE)) {
     x_est1 <- rowWeightedVars(x, w = w, na.rm = TRUE, useNames = useNames)
     print(x_est1)
     x_est2 <- colWeightedVars(t(x), w = w, na.rm = TRUE, useNames = useNames)
@@ -210,7 +180,7 @@ for (setDimnames in c(TRUE, FALSE)) {
   if (setDimnames) dimnames(x) <- dimnames
   else dimnames(x) <- NULL    
   # Check names attribute
-  for (useNames in c(if (!matrixStats:::isUseNamesNADefunct()) NA, TRUE, FALSE)) {
+  for (useNames in c(TRUE, FALSE)) {
     x_est1 <- rowWeightedVars(x, w = w, na.rm = TRUE, useNames = useNames)
     print(x_est1)
     x_est2 <- colWeightedVars(t(x), w = w, na.rm = TRUE, useNames = useNames)
@@ -235,12 +205,6 @@ x_est2 <- colWeightedSds(t(x), w = w, na.rm = FALSE, useNames = FALSE)
 stopifnot(all.equal(x_est1, x_est0))
 stopifnot(all.equal(x_est2, x_est0))
 x_est0 <- rowSds(x[, (w == 1), drop = FALSE], na.rm = FALSE, useNames = TRUE)
-if (!matrixStats:::isUseNamesNADefunct()) {
-  x_est1 <- rowWeightedSds(x, w = w, na.rm = FALSE, useNames = NA)
-  x_est2 <- colWeightedSds(t(x), w = w, na.rm = FALSE, useNames = NA)
-  stopifnot(all.equal(x_est1, x_est0))
-  stopifnot(all.equal(x_est2, x_est0))
-}  
 x_est1 <- rowWeightedSds(x, w = w, na.rm = FALSE, useNames = TRUE)
 x_est2 <- colWeightedSds(t(x), w = w, na.rm = FALSE, useNames = TRUE)
 stopifnot(all.equal(x_est1, x_est0))
@@ -263,12 +227,6 @@ x_est2 <- colWeightedMads(t(x), w = w, useNames = FALSE)
 stopifnot(all.equal(x_est1, x_est0))
 stopifnot(all.equal(x_est2, x_est0))
 x_est0 <- rowMads(x[, (w == 1), drop = FALSE], useNames = TRUE)
-if (!matrixStats:::isUseNamesNADefunct()) {
-  x_est1 <- rowWeightedMads(x, w = w, useNames = NA)
-  x_est2 <- colWeightedMads(t(x), w = w, useNames = NA)
-  stopifnot(all.equal(x_est1, x_est0))
-  stopifnot(all.equal(x_est2, x_est0))
-}
 x_est1 <- rowWeightedMads(x, w = w, useNames = TRUE)
 x_est2 <- colWeightedMads(t(x), w = w, useNames = TRUE)
 stopifnot(all.equal(x_est1, x_est0))